requirements.txt
langgraph>=0.3.34
langchain>=0.3.24
langchain-openai>=0.3.14
langchain-core>=0.3.56
pandas>=2.2.3
python-dotenv>=1.1.0
openinference-instrumentation-langchain>=0.1.0
arize-phoenix-otel>=0.9.2
pydantic>=2.11.3
